backend_flask/requirements.txt
flask
flask-cors
python-dotenv
requests
pycryptodome